requirements.txt
Flask
psycopg2-binary
flask_sqlalchemy
alembic
openai
python-dotenv